requirements.txt
asgiref==3.5.2
attrs==22.1.0
black==22.6.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
coverage==6.4.1
Django==4.0.5
django-countries==7.3.2
django-debug-toolbar==3.5.0
django-elasticsearch-dsl==7.2.2
django-js-asset==2.0.0
django-mptt==0.13.4
elasticsearch==7.17.6
elasticsearch-dsl==7.4.0
factory-boy==3.2.1
Faker==14.2.0
flake8==4.0.1
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
isort==5.10.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
Pillow==9.1.1
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.9
pytest==7.1.3
pytest-django==4.5.2
pytest-factoryboy==2.5.0
python-dateutil==2.8.2
python-decouple==3.6
requests==2.28.0
six==1.16.0
sqlparse==0.4.2
stripe==3.4.0
stripe3==0.0.0
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
